+++ activerecord/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Rails 5.2.2.1 (March 11, 2019) ##
+
+*   No changes.
+
+
 ## Rails 5.2.2 (December 04, 2018) ##
 
 *   Do not ignore the scoping with query methods in the scope block.